# Conflicts:
#	target/maven-archiver/pom.properties
#	target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst
#	target/maven-status/maven-compiler-plugin/testCompile/default-testCompile/createdFiles.lst
#	target/surefire-reports/ColumnTest.txt
#	target/surefire-reports/DataframeTest.txt
#	target/surefire-reports/TEST-ColumnTest.xml
#	target/surefire-reports/TEST-DataframeTest.xml

+++ src/main/java/Statistics.java
@@ -1,11 +1,12 @@
 public class Statistics {
+    
     public static float mean(Column c) throws TypeMismatchException{
         float somme = 0;
         if(c.getType() != Types.STRING){
             for(int i = 0; i < c.length(); i++){
-                somme += (float)c.getValue(i);
+                somme += (int)c.getValue(i);
             }
-            somme /= c.length();
+            somme /= (float)c.length();
         }else{
             throw new TypeMismatchException();
         }
